chore: move exceptions
